test: more scenarios for actions
