+++ src/ast/basic/semicolumn_term.rs
@@ -2,5 +2,6 @@ use ast_term_derive::ASTTerm;
 
 #[derive(ASTTerm, PartialEq, Eq, Debug, Clone)]
 pub struct SemicolumnTerm {
+    #[prop]
     location: Location,
 }
