+++ src/parser/single_module/tests/mod.rs
@@ -1,3 +1,3 @@
-mod actions;
+mod actions_and_functions;
 mod basic_cdl;
 mod imports;
